frontend/requirements.txt
beautifulsoup4==4.13.3
dash==2.18.2
dash_mantine_components==0.15.3
python-dotenv==1.0.1
requests==2.32.3
dash-iconify==0.1.2
flask-login==0.6.3
flask-sqlalchemy==3.1.1
